Make Seperate components for the tabs
